requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
PyYAML>=6.0.1
jinja2>=3.1.2
black>=23.10.0
requests>=2.31.0
python-dotenv>=1.0.0
openapi-spec-validator>=0.7.1 